+++ FFKit/gui/views/UIView+FFKit.h
@@ -32,7 +32,7 @@
  
     @see setX:adjustingWidth
 */
-@property (nonatomic, assign) float x;
+@property (nonatomic, assign) CGFloat x;
 
 /** Sets the x coordinate of the view.
  
@@ -42,13 +42,13 @@
  
     @see setX
 */
-- (void) setX: (float) newX adjustingWidth: (BOOL) adjustWidth;
+- (void) setX: (CGFloat) newX adjustingWidth: (BOOL) adjustWidth;
 
 /**	Gets or sets the y coordinate of the view.
     
     @see setY:adjustingHeight
 */
-@property (nonatomic, assign) float y;
+@property (nonatomic, assign) CGFloat y;
 
 /** Sets the y coordinate of the view.
  
@@ -58,25 +58,25 @@
  
     @see setY
 */
-- (void) setY: (float) newY adjustingHeight: (BOOL) adjustHeight;
+- (void) setY: (CGFloat) newY adjustingHeight: (BOOL) adjustHeight;
 
 /**	Gets or sets the width of the view. 
     
     @see setHeight
 */
-@property (nonatomic, assign) float width;
+@property (nonatomic, assign) CGFloat width;
 
 /**	Gets or sets the height of the view. 
  
     @see setWidth
 */
-@property (nonatomic, assign) float height;
+@property (nonatomic, assign) CGFloat height;
 
 /**	Gets or sets the x coordinate of the view's right edge. 
     
     @see setRight:adjustingWidth
 */
-@property (nonatomic, assign) float right;
+@property (nonatomic, assign) CGFloat right;
 
 /** Sets the x coordinate of the view's right edge.
  
@@ -86,13 +86,13 @@
  
     @see setRight
 */
-- (void) setRight: (float) newRight adjustingWidth: (BOOL) adjustWidth;
+- (void) setRight: (CGFloat) newRight adjustingWidth: (BOOL) adjustWidth;
 
 /**	Gets or sets the y coordinate of the view's bottom edge.
     
     @see setBottom:adjustingHeight
 */
-@property (nonatomic, assign) float bottom;
+@property (nonatomic, assign) CGFloat bottom;
 
 /** Sets the y coordinate of the view's bottom edge.
  
@@ -102,7 +102,7 @@
  
     @see setBottom
 */
-- (void) setBottom: (float) newBottom adjustingHeight: (BOOL) adjustHeight;
+- (void) setBottom: (CGFloat) newBottom adjustingHeight: (BOOL) adjustHeight;
 
 /** Changes the view's size and centres it within its superview. */
 - (void) centerWithSize: (CGSize) newSize;
@@ -116,13 +116,13 @@
  
     @see setCenter
 */
-@property (nonatomic, assign) float centerX;
+@property (nonatomic, assign) CGFloat centerX;
 
 /**	Gets or sets the center x of the view.
  
     @see setCenter
 */
-@property (nonatomic, assign) float centerY;
+@property (nonatomic, assign) CGFloat centerY;
 
 /**	Gets or sets whether or not the view is visible. */
 @property (nonatomic, assign) BOOL visible;
